content/media/webaudio/test/test_bug867174.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Crashtest for bug 867174</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SpecialPowers.setBoolPref("media.webaudio.enabled", true);
SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  var ctx = new AudioContext();

  var source = ctx.createBufferSource();
  var buffer = ctx.createBuffer(2, 2048, 8000);
  source.playbackRate.setTargetValueAtTime(0, 2, 3);
  var sp = ctx.createScriptProcessor();
  source.connect(sp);
  sp.connect(ctx.destination);
  source.start(0);

  sp.onaudioprocess = function(e) {
    // Now set the buffer
    source.buffer = buffer;

    ok(true, "We did not crash.");
    sp.onaudioprocess = null;
    SpecialPowers.clearUserPref("media.webaudio.enabled");
    SimpleTest.finish();
  };
});


</script>
</pre>
</body>
</html>
